Update segment.py and postprocess.py to detect.py and segment.py
